+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next release
 
 - Mark gems use for testing as development dependencies
+- Added `send_file`, `send_image`, `delete_file`, `delete_image`
 
 ## Oct 19th, 2019 - 1.0.0